pyproject.toml
[tool.poetry]
name = "arise"
version = "0.1.2"
description = "A Python project initializer designed to streamline the setup process for new Python projects. Arise simplifies the creation of project structures, dependency management, and configuration, allowing developers to focus on writing code. Whether you’re starting a new application, library, or script, Arise helps you get up and running quickly with a solid foundation."
authors = ["abdelmalek0 <djamaa.abdelmalek98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pygithub = "^2.4.0"
python-dotenv = "^1.0.1"
icecream = "^2.1.3"
rich = "^13.8.1"
simple-term-menu = "^1.6.4"
toml = "^0.10.2"

[tool.poetry.scripts]
arise = "arise.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
